Move directive parsing code to seperate module
