+++ tests/test_renderers/test_roles_directives/test_directive_options_error_block_style_text0_.xml
@@ -1,8 +1,8 @@
 <document source="">
     <system_message level="3" line="0" source="" type="ERROR">
         <paragraph>
-            Directive 'restructuredtext-test-directive' options:
-            Invalid YAML: mapping values are not allowed here
+            Directive 'restructuredtext-test-directive':
+            Invalid options YAML: mapping values are not allowed here
               in "<unicode string>", line 2, column 8:
                 option2: b
                        ^
